gpu/capture: lift the last symbols the collection seam crosses
Four more definitions both halves of gpu_capture.cpp reach, merged into the existing
internal header. Iterative by nature: each promotion changes every region index, so the
seam has to be re-derived and the next crossing surfaces. #3526 made merging possible;
before it, each round would have replaced the header.
